requirements.txt
# Core AI / agents
crewai>=0.28.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# App
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
